End-to-End-Projects/End-to-end-Source-Code-Analysis-Generative-AI/requirements.txt
langchain
langchain-openai
langchain-community
langchain-core
langchain-chroma
langchain-text-splitters
chromadb
flask
gitpython
python-dotenv
tiktoken
